requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
billiard==3.5.0.3
boto3==1.5.19
botocore==1.8.33
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
Django==2.0
django-appconf==1.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
django-ses-boto3==0.1.0
django-smtp-ssl==1.0
django-storages==1.6.5
djangorestframework==3.7.3
docutils==0.14
geoip2==2.6.0
idna==2.6
jmespath==0.9.3
kombu==4.1.0
Markdown==2.6.10
maxminddb==1.3.0
olefile==0.44
Pillow==4.3.0
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
urllib3==1.22
vine==1.1.4
